Added user and role features
updated it so when we create a student or faculty, it creates a user for that individual, and defaults the username and password to their email. Users who are logged in can now change their passwords. When users log in, they are routed to the appropriate module (Laura's logic). I updated the verbiage for the unauthorized and not found pages.